app/src/androidTest/java/com/ashutosh1234ojha/unittestapp/ui/ShoppingFragmentTest.kt
package com.ashutosh1234ojha.unittestapp.ui

import android.util.Log
import androidx.fragment.app.Fragment
import androidx.navigation.NavController
import androidx.navigation.Navigation
import androidx.test.espresso.Espresso.onView
import androidx.test.espresso.action.ViewActions.click
import androidx.test.espresso.matcher.ViewMatchers.withId
import androidx.test.filters.MediumTest
import com.ashutosh1234ojha.unittestapp.R
import com.ashutosh1234ojha.unittestapp.launchFragmentInHiltContainer
import dagger.hilt.android.AndroidEntryPoint
import dagger.hilt.android.testing.HiltAndroidRule
import dagger.hilt.android.testing.HiltAndroidTest
import kotlinx.coroutines.ExperimentalCoroutinesApi
import org.junit.Assert.*
import org.junit.Before
import org.junit.Rule
import org.junit.Test
import org.mockito.Mockito.mock
import org.mockito.Mockito.verify

/**
 * Created by Ashutosh Ojha on 13,November,2021
 */
@MediumTest
@HiltAndroidTest
@ExperimentalCoroutinesApi
class ShoppingFragmentTest {

    @get:Rule
    var hiltRule = HiltAndroidRule(this)

    @Before
    fun setUp() {
        hiltRule.inject()
    }

    @Test
    fun clickAddShoppingItemButton_navigateToAddShoppingItemFragment() {
        val navigationController = mock(NavController::class.java)
        launchFragmentInHiltContainer<ShoppingFragment> {
            Navigation.setViewNavController(requireView(), navigationController)
        }

        onView(withId(R.id.fabAddShoppingItem)).perform(click())

        verify(navigationController).navigate(R.id.action_shoppingFragment_to_addShoppingItemsFragment)
    }

}